[wmst] Fix bad allocation exception with high resolution tiled servers (fixes #30634)

+++ src/providers/wms/qgswmsprovider.cpp
@@ -660,7 +660,7 @@ QImage *QgsWmsProvider::draw( QgsRectangle const &viewExtent, int pixelWidth, in
 
       // if we know both source and output DPI, let's scale the tiles
       if ( mDpi != -1 && mTileLayer->dpi != -1 )
-        vres *= mDpi / mTileLayer->dpi;
+        vres *= static_cast<double>( mDpi ) / mTileLayer->dpi;
 
       // find nearest resolution
       tm = mTileMatrixSet->findNearestResolution( vres );
